==> src/jupyter/1.Export lookup tables and tpl files from excel.ipynb <==
### Step 1: Create swb input .txt and .tpl files from Excel spreadsheets

# 8/19/2020
# MNielsen
# Ran 8/3/2020 with MOREPARS files
# 9/9/202 re-ran after fixing MOREPARS files

import os
import numpy as np
import pandas as pd

outdir = r'E:\UMID_Data\0188_offline\modeldev\PEST\PEST_FILE_UPDATES'

# Location of lookup table spreadsheets:

pardir = r'E:\UMID_Data\0188_offline\modeldev\PARS\TABLES'

landuse_version = 'YETI_LU_lookup_table_MOREPARS'
irrigation_version = 'YETI_IRR_lookup_table_MOREPARS'

lufile = landuse_version + '.xlsx'
irrfile = irrigation_version + '.xlsx'

# Read landuse table
infile = os.path.join(pardir,lufile)
lutable = pd.read_excel(infile, 'TextExport')
tpltable1 = pd.read_excel(infile, 'TPL_Export')
del lutable['Unnamed: 0']
del tpltable1['Unnamed: 0']

# Read irrigation table
infile = os.path.join(pardir,irrfile)
irrtable = pd.read_excel(infile, 'TextExport')
tpltable2 = pd.read_excel(infile, 'TPL_Export')
#del irrtable['Unnamed: 0']
#del tpltable2['Unnamed: 0']

#### .txt and .tpl files:

# LU .txt file:
luoutfile = landuse_version + '.txt'  # filename
lupath = os.path.join(outdir, luoutfile)
lutable.to_csv(lupath, sep='\t')  # write file
# IRR .txt file:
irroutfile = irrigation_version + '.txt'   # filename
irrpath = os.path.join(outdir, irroutfile)
irrtable.to_csv(irrpath, sep = '\t')  # write file

# LU .tpl table:
lutplfile = landuse_version + '.tpl'   # filename
tpl1path = os.path.join(outdir, lutplfile)
tpltable1.to_csv(tpl1path, sep='\t')   # write file
# IRR .tpl table:
irrtplfile = irrigation_version + '.tpl'   # filename
tpl2path = os.path.join(outdir, irrtplfile)
tpltable2.to_csv(tpl2path, sep='\t')   # write file

# INSERT the "ptf ~" text in the .tpl files...

def line_prepender(filename, line):
    with open(filename, 'r+') as f:
        content = f.read()
        f.seek(0, 0)
        f.write(line.rstrip('\r\n') + '\n' + content)

line_prepender(tpl1path,'ptf ~')  # add to LU .tpl
line_prepender(tpl2path,'ptf ~')  # add to IRR .tpl
