==> src/jupyter/1b-Optional.Prep SW data calculations for PEST input.ipynb <==
### Make a CSV file of the SW observations with variances
#### Adjust initial values and weights based on COMMENT field (more weights work in pyemu notebook)
### Only needs to be run if new SW observations are added.

import os
import pandas as pd
import numpy as np

obspath = 'E:\\UMID_Data\\0188_offline\\modeldev\\OBS\\Baseflow_Separation_Results'
outpath = 'E:\\UMID_Data\\0188_offline\\modeldev\\OBS'

file1 = 'ANNUAL_RO_BF_OBS_TABLE.csv'
file2 = 'MONTHLY_RO_BF_OBS_TABLE.csv'  # This is ALL the years, not just the subset used in PEST


anndat = pd.read_csv(os.path.join(obspath, file1))
mondat = pd.read_csv(os.path.join(obspath, file2))

anndat.head()

## Variance is the square of the st dev:

alldat = anndat.append(mondat).reset_index()
alldat.info()

# First get rid of -9999 values AND losing reaches:
alldat = alldat.loc[alldat.OBSVAL > 0]

# Then adjust st dev if it is zero:
alldat.loc[alldat['obs_stdev'] == 0, 'obs_stdev'] = alldat['OBSVAL'] / 2

alldat['Variance'] = alldat['obs_stdev']**2

alldat.head()

tmp = alldat.groupby('OBS_group').describe().unstack(1).reset_index()
obsdat = tmp.loc[tmp['level_0'] == 'OBSVAL']
vardat = tmp.loc[tmp['level_0'] == 'Variance']

vardat.pivot(index='OBS_group',columns='level_1',values=0)

alldat_out = alldat[['IDX2','Comment','OBSVAL','OBS_group','Variance']]

outfile = 'OBSERVATIONS_DATA_BF_RO_wVARIANCES.csv'
alldat_out.to_csv(os.path.join(outpath, outfile))
