==> src/jupyter/2.Create OBS instruction file-MAP obs file.ipynb <==
# CREATE .ins File from the Observation Data file after running SWB:

# For MAP - NOT using a .csv file as the observation file (using .obs file)

# October, 2019
# Revised NOV 26/2019
# Revised JAN 30 2020
# M Nielsen

# 1/30/2020 ADDED ALL ET OBSERVATIONS.

# Initially Testing with a shortened number of observations
# Need to copy .csv file from YETI over to P: to make this work

# 7/15/2020 - re-did sorting so the .ins and .obs files are sorted the same.
# 8/10/2020 - used this to make an .ins file for the SW observations


### Run this when observations are changed, for whichever .obs file needs it.

import os
import pandas as pd
import numpy as np

obspath = 'E:\\UMID_Data\\0188_offline\\modeldev\\OBS'
#obsfile = 'swbstats_simulated_obs_et_ALL_sorted.obs'   # This one is sorted by ObsName

obsfile = 'swbstats_simulated_obs_SW.obs'   # This one is sorted by ObsName

# This file is just the SW observations (runoff and baseflow)

df = pd.read_csv(os.path.join(obspath, obsfile), sep=' ')

print (df.head())
print (df.info())

#del df['Unnamed: 0']  # --> Use if necessary.
df['Col1'] = 'l1'   # First character of the .ins line for each entry
df['insName'] = '~' + df['ObsName'] + '~'  # Second part of each .ins line
df['insOBS'] = '!' + df['ObsName'] + '!'  # OBS value part of each .ins line
print (df.tail())

print (df.info())

print (df.head())

# Clean and create output:
del df['ObsName']
#del df['Value']
del df['mean_swb']
#del df['Year1']
#del df['Year0']
print (df.head())

#outname = 'MAP_simulated_et_obs_ALL_ET_sorted.ins'
outname = 'MAP_simulated_obs_SW.ins'

outfile = os.path.join(obspath, outname)   # Can replace with /modeldev/PEST/PEST_FILE_UPDATES
df.to_csv(outfile, sep=' ', index=False, header=False)

print (df.head())

# INSERT the "pif ~" text in the file...

def line_prepender(filename, line):
    with open(filename, 'r+') as f:
        content = f.read()
        f.seek(0, 0)
        f.write(line.rstrip('\r\n') + '\n' + content)

line_prepender(outfile, 'l1')
line_prepender(outfile,'pif ~')
