==> src/jupyter/3.Parameter_Info_MAP-updated.ipynb <==
## Notebook to get parameter info from LOOKUP TABLES for MAP (Landuse and Irrigation)
### Script combines all parameters together and makes a .csv file with upper and lower bounds to use with pyemu

# Most recent run:
# 1/29/2020
# 8/10/2020
# 9/3/2020 (LU), 9/9/2020 (IRR)

import os
import numpy as np
import time
import pandas as pd
import shutil

### Set path and version for input Excel sheet:

# Location of lookup table spreadsheets:

pardir = r'E:\UMID_Data\0188_offline\modeldev\PARS\TABLES'

#landuse_version = 'YETI_Landuse_lookup_table_V4'
#irrigation_version = 'YETI_Irrigation_lookup_table_V2'
landuse_version = 'YETI_LU_lookup_table_MOREPARS'
irrigation_version = 'YETI_IRR_lookup_table_MOREPARS'

lufile = landuse_version + '.xlsx'
irrfile = irrigation_version + '.xlsx'

### Goal is to a) read excel file and find parameters, b) make a dataframe with PARUBND, PARLBND, etc., 
### c) get data from a second excel file (irrigation), and do the same.

### Starting with LU Table: Uses the BLUE colored tabs for Parnme and ParVal1

# Find starting values for each of these - read in two long lists and delete "--"
infile = os.path.join(pardir,lufile)

pnme = pd.read_excel(infile, 'Parnme', skiprows=3, na_values='--')
pval = pd.read_excel(infile, 'ParVal1', skiprows=3, na_values='--')
del pnme['Unnamed: 0']
del pval['Unnamed: 0']
#print (pval.info())
#print (pnme.info())

# Get the ParName list
test = pnme[~pnme.applymap(np.isreal).all(1)]
# This reduces the columns to ones that have text values:
test2 = test.select_dtypes(include='object')
print (test2.info())
test5 = pd.melt(test2, id_vars='description', value_name='ParName').dropna()
test6 = test5.apply(lambda x: x.str.strip('~'))
test7 = test6.apply(lambda x: x.str.strip()).reset_index()

# Create index location for the values:
test7['Index_Loc'] = test7['description'] + ":" + test7['variable']
test8 = test7.sort_values('Index_Loc')
print (test8.tail())

# Get the ParVal list
t3 = pval[~pval.applymap(np.isreal).all(1)]
t4 = t3.select_dtypes(include=['float64','object'])
#print (test2.info())
t5 = pd.melt(t4, id_vars='description', value_name='ParVal').dropna()
t5 = t5.applymap(lambda x: str(x).strip() if len(str(x).strip()) else None)  # Important - stripped off leading spaces

# Create index location for the values:
t5['Index_Loc'] = t5['description'] + ":" + t5['variable']
t6 = t5[['Index_Loc','ParVal']].sort_values('Index_Loc')
t6.ParVal = t6.ParVal.astype(np.float64)
print (t6.head())

# Join these together, remove the many duplicated items:
parvals = pd.merge(test8, t6, how='inner', on='Index_Loc', 
                   right_index=False, left_index=False, copy=False).drop_duplicates(['ParName','ParVal'])
print (parvals.info())

# Look for duplicated ParName with different ParVals - fix in spreadsheet:

# ADD CODE HERE to find items, fix in spreadsheet until they are gone.

mask = parvals.ParName.duplicated(keep=False)
dupes = (parvals[mask]).sort_values(['ParName'])
mask2 = dupes.duplicated(['ParName','ParVal'], keep=False)
# Want this to be empty:
realdupes = (dupes[~mask2])
realdupes.info()

realdupes.to_csv('landuse_pars_to_clean_up.csv')

### REPEAT for the Irrigation table:

# Find starting values for each of these - read in two long lists and delete "--"
infile = os.path.join(pardir,irrfile)

pnme2 = pd.read_excel(infile, 'Parnme', skiprows=2, na_values='--')
pval2 = pd.read_excel(infile, 'ParVal1', skiprows=2, na_values='--')
del pnme2['Unnamed: 0']
del pval2['Unnamed: 0']
#print (pval.info())

# Get the ParName list
irr1 = pnme2[~pnme2.applymap(np.isreal).all(1)]
irr2 = irr1.select_dtypes(include='object')
#print (test2)
irr3 = pd.melt(irr2, id_vars='description', value_name='ParName').dropna()
irr4 = irr3.apply(lambda x: x.str.strip('~'))
irr5 = irr4.apply(lambda x: x.str.strip()).reset_index()

# Create index location for the values:
irr5['Index_Loc'] = irr5['description'] + ":" + irr5['variable']
print (irr5.tail())

# Get the ParVal list
t10 = pval2[~pval2.applymap(np.isreal).all(1)]
t11 = t10.select_dtypes(include=['float64','object'])
#print (test2)
t12 = pd.melt(t11, id_vars='description', value_name='ParVal').dropna()
t12 = t12.applymap(lambda x: str(x).strip() if len(str(x).strip()) else None)  # Important - stripped off leading spaces

# Create index location for the values:
t12['Index_Loc'] = t12['description'] + ":" + t12['variable']
t13 = t12[['Index_Loc','ParVal']].copy()
t13.ParVal = t13.ParVal.astype(np.float64)
print (t13.info())

# Join these together:
parvals_irr = pd.merge(irr5, t13, how='outer', on='Index_Loc').drop_duplicates(['ParName','ParVal'])
print (parvals_irr.info())

# Look for duplicated ParName with different ParVals - fix in spreadsheet:

# ADD CODE HERE to find items, fix in spreadsheet until they are gone.

mask = parvals_irr.ParName.duplicated(keep=False)
dupes = (parvals_irr[mask]).sort_values(['ParName'])
mask2 = dupes.duplicated(['ParName','ParVal'], keep=False)
# Want this to be empty:
realdupes = (dupes[~mask2])
realdupes.info()

realdupes.to_csv('irrigation_pars_to_clean_up.csv')

## Join LU and IRR pars together:

pars_all = parvals.append(parvals_irr, ignore_index=False).reset_index()
print (pars_all.info())

# Find any duplicates:
#mask = df.B.duplicated(keep=False)
mask = pars_all.ParName.duplicated(keep=False)
print (pars_all[mask])

# If needed, remove duplicates
pars_all = pars_all.drop_duplicates(['ParName']).copy()
pars_all['ParName'] = pars_all['ParName'].str.lower()
pars_all.info()

print (pars_all.tail(10))

print (pars_all.ParName.unique())

### Need to set parameter upper and lower bounds, par groups

### Par Groups:

# First, the PAR GROUPS:
pars_all['ParGroup'] = pars_all['ParName'].str.split("_").str[-1]
print (pars_all.tail())

print (pars_all['ParGroup'].unique())
print (pars_all.info())

### Upper and Lower Bounds:

# Set Upper and Lower bounds for Curve Numbers:
# Arbitrary range of 30 - 99 in total. 
pars_all['ParUBND'] = -99
pars_all['ParLBND'] = -999
maxCN = 99.0
minCN = 30.0
# Upper BND =  (ParVal + (MaxCN - ParVal)/2)
pars_all['tmp'] = maxCN - pars_all['ParVal']
pars_all['tmp'] = pars_all['tmp'] / 2
pars_all['tmp'] = pars_all.ParVal + pars_all.tmp
pars_all.loc[pars_all.ParGroup == 'cn', 'ParUBND'] = pars_all['tmp']
# Lower BND =  (ParVal - (ParVal - MinCN)/2)
pars_all['tmp'] = pars_all['ParVal'] - minCN
pars_all['tmp'] = pars_all['tmp'] / 2
pars_all['tmp'] = pars_all.ParVal - pars_all.tmp
pars_all.loc[pars_all.ParGroup == 'cn', 'ParLBND'] = pars_all['tmp']

print (pars_all.head())

# Set upper and lower bounds for ROOTING ZONE DEPTHS
# Just use a multiplier, nothing fancy:
pars_all.loc[pars_all.ParGroup == 'rz', 'ParUBND'] = pars_all['ParVal'] * 1.5
pars_all.loc[pars_all.ParGroup == 'rz', 'ParLBND'] = pars_all['ParVal'] * 0.5

print (pars_all.head())

# Set for PLANT HEIGHTS:
# Use same thing: simple multipliers
pars_all.loc[pars_all.ParGroup == 'height', 'ParUBND'] = pars_all['ParVal'] * 1.5
pars_all.loc[pars_all.ParGroup == 'height', 'ParLBND'] = pars_all['ParVal'] * 0.5

# Set for Kcb-Mid values:
# Use same thing: simple multipliers
pars_all.loc[pars_all.ParGroup == 'kcb-mid', 'ParUBND'] = pars_all['ParVal'] * 1.25
pars_all.loc[pars_all.ParGroup == 'kcb-mid', 'ParLBND'] = pars_all['ParVal'] * 0.75

# Set for REW and TEW settings:
pars_all.loc[pars_all.ParGroup == 'rew', 'ParUBND'] = pars_all['ParVal'] * 1.5
pars_all.loc[pars_all.ParGroup == 'rew', 'ParLBND'] = pars_all['ParVal'] * 0.1
pars_all.loc[pars_all.ParGroup == 'tew', 'ParUBND'] = pars_all['ParVal'] * 1.5
pars_all.loc[pars_all.ParGroup == 'tew', 'ParLBND'] = pars_all['ParVal'] * 0.1

print (pars_all.tail())

# Max Net Infiltration: same thing, simple multipliers:
pars_all.loc[pars_all.ParGroup == 'mni', 'ParUBND'] = pars_all['ParVal'] * 2
pars_all.loc[pars_all.ParGroup == 'mni', 'ParLBND'] = pars_all['ParVal'] * 0.25

# Find any that haven't been set yet:

pars_all.loc[pars_all['ParUBND'] < 0]

### Adjustable:

pars_all['Adjust'] = 'YES'
print (pars_all.info())

### Clean up for exporting:

pars_exp = pars_all[['ParName','ParGroup','ParVal','ParUBND','ParLBND','Adjust']].copy()

outdir = r"E:\UMID_Data\0188_offline\modeldev\PARS"
pars_exp.to_csv(os.path.join(outdir,'MAP_parlist_MOREPARS.csv'))
